refactor(application_controller): Adding Api::V1::BaseController like @mrbongiolo said
